+++ meeshkan/__main__.py
@@ -26,7 +26,7 @@ from .core.api import Api
 from .core.service import Service
 from .core.logger import setup_logging, remove_non_file_handlers
 from .core.job import Job
-from .utils import save_token, get_auth, _get_api, _build_cloud_client
+from .__utils__ import save_token, get_auth, _get_api, _build_cloud_client
 from .start import start_agent
 
 LOGGER = None
